Refactor MLIR lowering passes into separate files
